Add files via upload
Add SQL dump of Chinook database

This commit includes a SQL file containing the full schema and data of the Chinook database. The file can be used to recreate the database schema and populate it with the sample data.

The Chinook database simulates a digital media store, including tables for artists, albums, customers, employees, invoices, tracks, and more. This file is intended for educational and demonstration purposes.